+++ scripts/dataset_builder/__init__.py
@@ -2,7 +2,7 @@
 from .downloader import fetch_all_sources, load_sources_config, download_file
 from .clustering import build_variant_graph, classify_cluster, parse_opencc_file, is_cjk, UnionFind
 from .kana import romaji_to_katakana, romaji_to_hiragana
-from .unihan import parse_unihan_readings, enrich_cluster
+from .unihan import parse_unihan_readings, parse_unihan_definitions, enrich_cluster
 
 __all__ = [
     "fetch_all_sources",
@@ -16,5 +16,6 @@ __all__ = [
     "romaji_to_katakana",
     "romaji_to_hiragana",
     "parse_unihan_readings",
+    "parse_unihan_definitions",
     "enrich_cluster",
 ]